requirements.txt
beautifulsoup4==4.12.2
google_api_python_client==2.106.0
google_auth_oauthlib==1.1.0
html2text==2020.1.16
protobuf==4.25.0
python-dotenv==1.0.0
python-telegram-bot==20.6
PyYAML==6.0.1
redis==5.0.1
Requests==2.31.0
selenium==4.15.2
zhipuai==1.0.7
python-telegram-bot[job-queue]==20.6